Refactor DataSourcesSection to extract DatasetItem
Extract the inline dataset rendering block into a distinct `DatasetItem` component, migrating all localized state mapping and necessary store access hooks to the child. This cleans up the massive `DataSourcesSection` function block and improves maintainability and encapsulation.
